pyproject.toml
[project]
name = "battleships"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.128.0",
  "pydantic>=2.12.5",
  "pydantic-settings>=2.12.0",
  "python-dotenv>=1.2.1",
  "uvicorn>=0.40.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["test"]

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "requests>=2.32.5",
    "ruff>=0.14.14",
    "ty>=0.0.14",
]

[tool.pyright]
include = ["src", "test"]
extraPaths = ["src"]

typeCheckingMode = "strict"

pythonVersion = "3.12"
venvPath = "."
venv = ".venv"

# Prefer explicit typing, but as warnings
reportMissingParameterType = "warning"
reportUnknownParameterType = "warning"
reportUnknownVariableType = "warning"

[tool.poe.tasks]
ruff = "ruff check ."
pyright = "pyright"
ty = "ty check"

check = ["ruff", "pyright", "ty"]
dev = "uvicorn main:app --app-dir src --host 127.0.0.1 --port 8000 --reload"
debug = "python src/debug/debugger.py"
